refactor: Header 구조 리펙토링(#10)
