+++ src/components/Header/components/HeaderTab/index.jsx
@@ -6,6 +6,27 @@ import Logo from './logo';
 import MenuBar from './menuBar';
 import OrnamentalSearchBar from './ornamentalSearchBar';
 
+export default function HeaderTab({ activeSearchTab }) {
+  return (
+    <Container>
+      <UpperContainer>
+        <Group>
+          <Logo />
+          <OrnamentalSearchBar onClick={activeSearchTab} />
+        </Group>
+        <Bookmark />
+      </UpperContainer>
+      <BottomContainer>
+        <MenuBar />
+      </BottomContainer>
+    </Container>
+  );
+}
+
+HeaderTab.propTypes = {
+  activeSearchTab: PropTypes.func.isRequired,
+};
+
 const Container = styled.div``;
 
 const UpperContainer = styled.div`
@@ -23,24 +44,3 @@ const Group = styled.div`
   align-items: center;
   gap: 30px;
 `;
-
-HeaderTab.propTypes = {
-  activeSearchTab: PropTypes.func.isRequired,
-};
-
-export default function HeaderTab({ activeSearchTab }) {
-  return (
-    <Container>
-      <UpperContainer>
-        <Group>
-          <Logo />
-          <OrnamentalSearchBar onClick={activeSearchTab} />
-        </Group>
-        <Bookmark />
-      </UpperContainer>
-      <BottomContainer>
-        <MenuBar />
-      </BottomContainer>
-    </Container>
-  );
-}
