+++ src/components/Header/components/HeaderTab/menuBar.jsx
@@ -1,46 +1,12 @@
 import styled from 'styled-components';
 import { useState, useCallback, useRef } from 'react';
+import { Link, useLocation } from 'react-router-dom';
 
 import { MenuIcon } from '@/common/icons';
-import { Link, useLocation } from 'react-router-dom';
 import { EVENT_PATH, OFFICIAL_PATH } from '@/constants/route';
 import CategoryOverlay from '../CategoryOverlay/index';
 import { FloatingMenu } from '@/common/components';
 import useTimer from '@/common/hooks/useTimer';
-
-const Container = styled.div`
-  position: relative;
-  color: #444444;
-  font-size: 14px;
-  font-weight: 700;
-  margin-top: 30px;
-`;
-
-const MenuItem = styled.div`
-  cursor: pointer;
-  display: inline-block;
-  text-align: center;
-  height: 35px;
-  user-select: none;
-  padding: 0px 23px;
-
-  border-bottom: ${({ theme, active }) =>
-    active ? `2px solid ${theme.colors.logo}` : 'none'};
-
-  &:hover {
-    border-bottom: ${({ theme }) => `2px solid ${theme.colors.border}`};
-  }
-`;
-
-const CategoryItem = styled(MenuItem)`
-  padding-left: 0;
-`;
-
-const MenuLink = styled(Link)`
-  display: block;
-  width: 100%;
-  height: 100%;
-`;
 
 export default function MenuBar() {
   const { pathname } = useLocation();
@@ -87,3 +53,37 @@ export default function MenuBar() {
     </>
   );
 }
+
+const Container = styled.div`
+  position: relative;
+  color: #444444;
+  font-size: 14px;
+  font-weight: 700;
+  margin-top: 30px;
+`;
+
+const MenuItem = styled.div`
+  cursor: pointer;
+  display: inline-block;
+  text-align: center;
+  height: 35px;
+  user-select: none;
+  padding: 0px 23px;
+
+  border-bottom: ${({ theme, active }) =>
+    active ? `2px solid ${theme.colors.logo}` : 'none'};
+
+  &:hover {
+    border-bottom: ${({ theme }) => `2px solid ${theme.colors.border}`};
+  }
+`;
+
+const CategoryItem = styled(MenuItem)`
+  padding-left: 0;
+`;
+
+const MenuLink = styled(Link)`
+  display: block;
+  width: 100%;
+  height: 100%;
+`;
