+++ src/components/Header/components/SearchTab/History/historyList.jsx
@@ -5,35 +5,6 @@ import styled from 'styled-components';
 import { XIcon } from '@/common/icons';
 import { ignorePrefix } from '@/utils/str';
 
-const Container = styled.ul`
-  font-size: 18px;
-  padding-top: 20px;
-  padding-bottom: 15px;
-`;
-
-const Item = styled.li`
-  display: flex;
-  justify-content: space-between;
-  padding: 7px 0px;
-  border-radius: 8px;
-  cursor: pointer;
-  &:hover {
-    background-color: #f5f6f7;
-  }
-`;
-
-const ItemTitle = styled.div`
-  flex-grow: 1;
-`;
-
-const DeleteButton = styled.button``;
-
-HistoryList.propTypes = {
-  history: PropTypes.array.isRequired,
-  removeHistory: PropTypes.func.isRequired,
-  addTag: PropTypes.func.isRequired,
-};
-
 /**
  * @param {{history: string[], removeHistory(index: number):void, addTag(tag): void}} props
  */
@@ -59,3 +30,32 @@ export default function HistoryList({ history, removeHistory, addTag }) {
     </Container>
   );
 }
+
+HistoryList.propTypes = {
+  history: PropTypes.array.isRequired,
+  removeHistory: PropTypes.func.isRequired,
+  addTag: PropTypes.func.isRequired,
+};
+
+const Container = styled.ul`
+  font-size: 18px;
+  padding-top: 20px;
+  padding-bottom: 15px;
+`;
+
+const Item = styled.li`
+  display: flex;
+  justify-content: space-between;
+  padding: 7px 0px;
+  border-radius: 8px;
+  cursor: pointer;
+  &:hover {
+    background-color: #f5f6f7;
+  }
+`;
+
+const ItemTitle = styled.div`
+  flex-grow: 1;
+`;
+
+const DeleteButton = styled.button``;
