+++ src/components/Header/components/SearchTab/index.jsx
@@ -6,27 +6,6 @@ import useTagSearch from '@/components/Header/hooks/useTagSearch';
 import Recommend from './recommend';
 import History from './History';
 
-const Container = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  background-color: ${({ theme }) => theme.backgroundColor.main};
-`;
-
-const Content = styled.div`
-  margin: 0 auto;
-  padding: 30px 0px;
-
-  ${({ theme }) => theme.media.desktop`
-    ${'width: ' + theme.widths.desktop + 'px'};
-  `};
-
-  ${({ theme }) => theme.media.tablet`
-    ${'width: ' + theme.widths.tablet + 'px'};
-  `};
-`;
-
 /**
  * @param {{afterSearch():void}} param0
  * @returns
@@ -51,3 +30,24 @@ export default function SearchTab({ afterSearch }) {
     </Container>
   );
 }
+
+const Container = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  background-color: ${({ theme }) => theme.backgroundColor.main};
+`;
+
+const Content = styled.div`
+  margin: 0 auto;
+  padding: 30px 0px;
+
+  ${({ theme }) => theme.media.desktop`
+    ${'width: ' + theme.widths.desktop + 'px'};
+  `};
+
+  ${({ theme }) => theme.media.tablet`
+    ${'width: ' + theme.widths.tablet + 'px'};
+  `};
+`;
