+++ src/pages/ClientOutlet.jsx
@@ -1,5 +1,5 @@
 import Banner from '@/components/Banner';
-import Header from '@/components/Header';
+import { Header } from '@/components/Header';
 import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
